+++ src/components/TextRenderers/__deprecated__/NASpan.tsx
