+++ src/components/TextRenderers/__deprecated__/SparkPercentageBar.tsx
