+++ src/components/TextRenderers/__deprecated__/index.ts
