+++ src/components/TextRenderers/__deprecated__/utils.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { NASpan } from "@text/__deprecated__/NASpan"
+import { NASpan } from "@components/TextRenderers/__deprecated__/NASpan"
 
 /* altValue defines what should be displayed if not n/a */
 export const resolveNAs = (value: string, altValue: any = null) => {
